Вынес структуру конфига в отдельный файл
