+++ go.mod
@@ -3,13 +3,14 @@ module tutor
 go 1.22
 
 require (
-	github.com/avast/retry-go v2.7.0+incompatible
+	github.com/go-telegram-bot-api/telegram-bot-api v4.6.4+incompatible
+	github.com/pkoukk/tiktoken-go v0.1.7
 	github.com/sashabaranov/go-openai v1.25.0
 )
 
 require (
 	github.com/dlclark/regexp2 v1.10.0 // indirect
 	github.com/google/uuid v1.3.0 // indirect
-	github.com/pkoukk/tiktoken-go v0.1.7 // indirect
 	github.com/stretchr/testify v1.9.0 // indirect
+	github.com/technoweenie/multipartstreamer v1.0.1 // indirect
 )
